Configure batch changes repo route from enterprise code
