+++ client/web/src/enterprise/repo/routes.tsx
@@ -1,7 +1,31 @@
+import React from 'react'
+
 import { RepoContainerRoute } from '../../repo/RepoContainer'
 import { RepoRevisionContainerRoute } from '../../repo/RepoRevisionContainer'
 import { repoContainerRoutes, repoRevisionContainerRoutes } from '../../repo/routes'
+import { lazyComponent } from '../../util/lazyComponent'
 
-export const enterpriseRepoContainerRoutes: readonly RepoContainerRoute[] = repoContainerRoutes
+const RepositoryGitDataContainer = lazyComponent(
+    () => import('../../repo/RepositoryGitDataContainer'),
+    'RepositoryGitDataContainer'
+)
+
+const RepositoryBatchChangesArea = lazyComponent(
+    () => import('../batches/repo/RepositoryBatchChangesArea'),
+    'RepositoryBatchChangesArea'
+)
+
+export const enterpriseRepoContainerRoutes: readonly RepoContainerRoute[] = [
+    ...repoContainerRoutes,
+
+    {
+        path: '/-/batch-changes',
+        render: context => (
+            <RepositoryGitDataContainer {...context} repoName={context.repo.name}>
+                <RepositoryBatchChangesArea {...context} />
+            </RepositoryGitDataContainer>
+        ),
+    },
+]
 
 export const enterpriseRepoRevisionContainerRoutes: readonly RepoRevisionContainerRoute[] = repoRevisionContainerRoutes
